+++ src/views/index.ts
@@ -1,5 +1,5 @@
-export * from "./ManageList";
+export * from "./authenticated/ManageList";
 export * from "./Home";
 export * from "./Layout";
-export * from "./List";
-export * from "./PageNotFound";
+export * from "./authenticated/List";
+export * from "./unauthenticated/PageNotFound";
